Remove edoc_doclet_default
edoc_doclet_default was refactored out from edoc_doclet when
edoc_doclet_chunks was created. edoc_doclet was supposed to just define
the behaviour and types, whereas the other modules the implementations.

A rebase mistake led to edoc_doclet and edoc_doclet_default having the
same contents. However, @garazdawi pointed out that edoc_doclet:run/2 is
part of the public, documented API, so it's better not to rename/split it.

This leaves us with the chunk doclet in edoc_doclet_chunks and the
default one in edoc_doclet.